requirements.txt
fastapi==0.95.2  # Ensure compatibility with the FastAPI version you're using
pydantic[email]  # For email validation with Pydantic
python-jose==3.3.0  # For JWT handling
passlib[bcrypt]==1.7.4  # For password hashing
motor==3.1.1  # For asynchronous MongoDB interaction
email-validator==1.3.1  # Required for Pydantic email validation
python-dotenv==1.0.0  # For loading environment variables from .env
uvicorn==0.22.0  # ASGI server for running the FastAPI app
